refactor(orchestration): remove probing facade exports
