+++ tests/orchestration/test_preparation.py
@@ -12,7 +12,7 @@ import pytest
 import frame_compare.analysis.cache_io as cache_io
 from frame_compare.errors import MetricsCalculationError
 from frame_compare.orchestration import preparation
-from frame_compare.orchestration.probing import load_clip_probe_cache
+from frame_compare.orchestration.probing.probe_cache import load_clip_probe_cache
 from frame_compare.orchestration.types import RunDependencies, RunRequest
 from frame_compare.services.alignment import CACHE_FILE_NAME
 from frame_compare.services.errors import AudioAlignmentError
